Refactor SiteDB into two parts to make it easy to mock
